+++ CHANGELOG.md
@@ -355,4 +355,6 @@ This update makes several changes to Predator that dramatically improve its proc
 
 *Release date to be determined*
 
-- Added the ability to set the units for speed in the OpenCV GPS dashcam video overlay stamp.
+- Refined OpenCV dashcam recording.
+    - Added the ability to set the units for speed in the OpenCV GPS dashcam video overlay stamp.
+    - Made the font color of the video overlay stamps configurable.